requirements.txt
pydantic
pytest
sqlmodel
python-dotenv